Memcached script Chose

+++ memcached.py
@@ -6,36 +6,52 @@
 
 import psutil, sys, time
 import subprocess
+import os
 
 
 class Memcached(object):
-    def __init__(self, processname):
+    def __init__(self, processname, pid_path):
         self.processname = processname
+        self.pid_path = pid_path
 
     def stop_memcacehd(self):
-        print "\033[32mStop Memcached...............\033[0m"
-        for process in psutil.process_iter():
-            if process.name() == self.processname:
-                process.kill()
+        if os.path.exists(self.pid_path):
+            print "\033[32mStop Memcached...............\033[0m"
+            for process in psutil.process_iter():
+                if process.name() == self.processname:
+                    process.kill()
+                    print "\033[32m----------Stop %s ProcessID=%s user=%s Successful----------\033[0m" % (
+                        process.name(), process.pid, process.username())
+            os.remove(self.pid_path)
+        else:
+            print "\033[32m-----------Service does not exist||Starting Memcacehd----------"
 
     def start_memcached(self):
         print "\033[32mStart Memcached.............\033[0m"
         memcached_bin_home = '/software/memcached/bin/memcached'
         for port in range(11211, 11215):
-            memcached_start_options = "%s -d -m 10 -u root -l 0.0.0.0 -p %s -c 512 -P /tmp/memcached.pid" % (
-                memcached_bin_home, port)
+            memcached_start_options = "%s -d -m 10 -u root -l 0.0.0.0 -p %s -c 512 -P %s" % (
+                memcached_bin_home, port, self.pid_path)
             subprocess.call(memcached_start_options, shell=True)
+            print "\033[32m----------Start At Port %s ---------\033[0m" % port
+            time.sleep(1)
+        if os.path.exists(self.pid_path):
+            print "\033[32mMemcached Start Successful\033[0m"
 
     def get_memcacehd_info(self):
-        for process in psutil.process_iter():
-            if process.name() == self.processname:
-                print "程序PID=>", process.pid, "\n", "启动路径=>", process.exe(), \
-                    "\n", "运行用户=>", process.username(), "\n", "运行状态=>", process.status()
+        if os.path.exists(self.pid_path):
+            for process in psutil.process_iter():
+                if process.name() == self.processname:
+                    print "程序PID=>", process.pid, "\n", "启动路径=>", process.exe(), \
+                        "\n", "运行用户=>", process.username(), "\n", "运行状态=>", process.status()
+        else:
+            print "\033[32mMemcached Is Don't Start\033[0m"
 
 
 if __name__ == '__main__':
-    memcacehd_process_name = 'memcached'
-    run = Memcached(memcacehd_process_name)
+    memcached_process_name = 'memcached'
+    memcached_pid_path = '/tmp/memcached.pid'
+    run = Memcached(memcached_process_name, memcached_pid_path)
     try:
         if sys.argv[1] == 'info':
             run.get_memcacehd_info()
